Added some basic coding to make the player character animate when moving. Added some code to display the second note graphic.
Rewrote the keypad script to fix a small error. You can now also input a maximum of 7 digits. If you press a key after reaching the cap it resets back to 0 inputs.
Remade the controls scene.
Fixed the keypad object in the level so it doesn't show on the other side of the wall, inside of the secret room.